requirements.txt
adb_shell[async]>=0.4.3
flask>=0.12.2
pyyaml>=3.12
pychromecast>=12.1.3
requests>=2.18.4
udi-interface>=3.0.51
wakeonlan>=2.1.0
ws4py>=0.5.1
